feat: sync authorizer openapi spec with api
Copy the api specification to a subpath of the settings for any
lmabda authorizers used.

The authorizer relies on a copy of the openapi specification to
determine various attributes of methods such as scopes and acr.

It thus needs a new copy of the api spec each time it is changed by the
api. To achieve this, copy the spec to a specific subpath of the
authorizer settings. A subpath ensures other mechanisms such as asFile
or config do not interfere with the copied spec.