feat: Message Transfer Agent components
Implements #1499

First implementation of inbound and outbound MTAs.

The outbound implementation mainly provides link and attribute
support to permit sending of emails.

The inbound implementation supports rulesets and IP filters. S3, lambda
and drop actions are supported.

Because inbound email support is only available in limited regions, the
implementation currently assumes that if S3 encryption is required,
AES will be used. KMS is more problematic as the framework support for
sharing KMS keys is not yet complete.