fix: permission on globaldb secopndary indexes
Where a globaldb is defined to have secondary indexes, linking to a
globaldb will also grant access to the secondard indexes.

This fix required some refactoring of the dynamodb policy support to
correctly support permissions on secondard indexes. The changes were
made to be backwards compatible, but this did result in a degree
of inconsistency in the ordering of parameters to some methods that
previously didn't accept a list of secondary indexes.