feat: baselinekey extensions and policy migration (#229)
* feat(baseline): add support for policy on keys

adds support for extension based policy definitions for the baseline key
component. This allows for ongoing configuration of the key policy to
align with the requirements of solution specific policies instead of
hamlet global policies

* refactor(baseline): migrate existing key policy

Moves the existing cmk key policies over to aws specific extensions. The
current policy is applied via masterdata

adds a new policy to allow S3 access to the CMK when delivering
inventory reports