fix: masterdata object validation errors
BREAKING CHANGE: In hamlet-io/engine#1548 the composite object model is being updated
to allow for validation of the structure - including errors when
additional parameters are present.

Testing of that has surfaced a number of incorrect objects in this
plugin that have gone unnoticed

Commit is a breaking change as - though no new functionality has been
implemented - the corrections will impact existing deployments (to
their originally intended state).